+++ src/Presentation/Nop.Web/Areas/Admin/Models/Customers/CustomerAttributeValueModel.cs
@@ -19,7 +19,7 @@ public partial record CustomerAttributeValueModel : BaseNopEntityModel, ILocaliz
 
     #region Properties
 
-    public int AttributeId { get; set; }
+    public long AttributeId { get; set; }
 
     [NopResourceDisplayName("Admin.Customers.CustomerAttributes.Values.Fields.Name")]
     public string Name { get; set; }
@@ -37,7 +37,7 @@ public partial record CustomerAttributeValueModel : BaseNopEntityModel, ILocaliz
 
 public partial record CustomerAttributeValueLocalizedModel : ILocalizedLocaleModel
 {
-    public int LanguageId { get; set; }
+    public long LanguageId { get; set; }
 
     [NopResourceDisplayName("Admin.Customers.CustomerAttributes.Values.Fields.Name")]
     public string Name { get; set; }
